+++ CHANGELOG.md
@@ -34,7 +34,7 @@
 * `cases` kwarg to `dvars_sensitivity` to allow specifying which cases to use for the sensitivity analysis ([GH-14](https://github.com/scottshambaugh/monaco/issues/14))
 * `cases` kwarg to `plot_2d_cov_ellipse`, `plot_integration_error`, and `plot_integration_convergence` to allow specifying which cases to use for the plot ([GH-14](https://github.com/scottshambaugh/monaco/issues/14))
 ### Changed    
-* Switch varstatbootstrap method from 'basic' to 'BCa'
+* Switch varstat bootstrap method from 'basic' to 'BCa'
 * `get_cases` function moved from `mc_plot` to `helper_functions`
 
 ## [0.14.1] - 2025-01-31